Fix framework and available expressions
**Description**

This commit adds changes fixes in the framework
code, and additional logging for debugging, in building
the domain, as well as the transfer function bitvector.

**Testing**

1. Generating the .ll file to run the pass on.

Input
=================
int main(int argc, char *argv[]) {
  int a, b, c, d, e, f;
  a = 50;
  b = argc + a;
  c = 96;
  e = b + c;
  if (a < b) {
    f = b - a;
    e = c * b;
  } else {
    f = b + a;
    e = c * b;
  }
  b = a - c;
  d = b + f;
  return 0;
}

Run the mem2reg pass

clang-12 -O0 -Xclang -disable-O0-optnone -emit-llvm -c test.c -o test.bc
opt-12 -passes=mem2reg test.bc -o testM2R.bc llvm-dis-12 testM2R.bc -o=AvailExpr.ll

2. Run the avail-expr pass

opt-12 -load build/lib/libDFA.so -avail-expr test/AvailExpr.ll -disable-output

+++ Assignment2-Dataflow_Analysis/include/dfa/MeetOp.h
@@ -35,7 +35,7 @@ public:
     }
 
     DomainVal_t result(LHS.size(), false);
-    for (int i = 0; i < LHS.size(); i++) {
+    for (long unsigned int i = 0; i < LHS.size(); i++) {
       result[i] = LHS[i] && RHS[i];
     }
 
@@ -70,7 +70,7 @@ public:
     }
 
     DomainVal_t result(LHS.size(), false);
-    for (int i = 0; i < LHS.size(); i++) {
+    for (long unsigned int i = 0; i < LHS.size(); i++) {
       result[i] = LHS[i] || RHS[i];
     }
